refactor: 지하철 구간 관리 입출력을 모두 controller에서 담당
