+++ src/main/java/subway/section/SectionController.java
@@ -1,6 +1,13 @@
 package subway.section;
 
-import subway.common.CommonService;
+import subway.line.Line;
+import subway.line.LineService;
+import subway.line.validation.CheckAlreadyRegisteredStation;
+import subway.line.validation.CheckNotExistStation;
+import subway.line.validation.CheckRightSectionNumber;
+import subway.main.SubwayController;
+import subway.station.Station;
+import subway.station.StationService;
 import subway.view.InputView;
 import subway.view.section.SectionManagementView;
 
@@ -17,7 +24,7 @@ public class SectionController {
 
         while (true) {
             SectionManagementView.showSectionManagementMenu();
-            char option = CommonService.selectOption(optionList, inputView);
+            char option = SubwayController.selectOption(optionList, inputView);
 
             if (option == GO_BACK) {
                 break;
@@ -40,22 +47,66 @@ public class SectionController {
     }
 
     private static boolean addNewSection(InputView inputView) {
-        SectionManagementView.askAddSectionLineName();
-        String lineName = inputView.lineName();
-        boolean success = SectionService.addSection(lineName, inputView);
-        if (success) {
-            SectionManagementView.addSectionComplete();
+        boolean success = false;
+        try {
+            Line line = getAddSectionLine(inputView);
+            Station station = getAddSectionStation(line, inputView);
+            int sectionOrder = getAddSectionOrder(line, inputView);
+            success = SectionService.addSection(line, station, sectionOrder);
+            if (success) {
+                SectionManagementView.addSectionComplete();
+            }
+        } catch (IllegalArgumentException e) {
+            System.out.println(e.getMessage());
         }
         return success;
     }
 
-    private static boolean deleteSection(InputView inputView) {
-        SectionManagementView.askDeleteSectionLineName();
+    private static Line getAddSectionLine(InputView inputView) {
+        SectionManagementView.askAddSectionLineName();
         String lineName = inputView.lineName();
-        boolean success = SectionService.deleteSection(lineName, inputView);
-        if (success) {
-            SectionManagementView.deleteSectionComplete();
+        return LineService.findLine(lineName);
+    }
+
+    private static Station getAddSectionStation(Line line, InputView inputView) {
+        SectionManagementView.askAddStationName();
+        String stationName = inputView.stationName();
+        CheckAlreadyRegisteredStation.validation(line, stationName);
+        return StationService.findStation(stationName);
+    }
+
+    private static int getAddSectionOrder(Line line, InputView inputView) {
+        SectionManagementView.askStationOrder();
+        String number = inputView.sectionNumber();
+        CheckRightSectionNumber.validation(line, number);
+        return Integer.parseInt(number);
+    }
+
+    private static boolean deleteSection(InputView inputView) {
+        boolean success = false;
+        try {
+            Line line = getDeleteSectionLine(inputView);
+            Station station = getDeleteSectionStation(line, inputView);
+            success = SectionService.deleteSection(line, station);
+            if (success) {
+                SectionManagementView.deleteSectionComplete();
+            }
+        } catch (IllegalArgumentException e) {
+            System.out.println(e.getMessage());
         }
         return success;
+    }
+
+    private static Line getDeleteSectionLine(InputView inputView) {
+        SectionManagementView.askDeleteSectionLineName();
+        String lineName = inputView.lineName();
+        return LineService.findLine(lineName);
+    }
+
+    private static Station getDeleteSectionStation(Line line, InputView inputView) {
+        SectionManagementView.askDeleteSectionStation();
+        String stationName = inputView.stationName();
+        CheckNotExistStation.validation(line, stationName);
+        return StationService.findStation(stationName);
     }
 }
